backend/requirements.txt
fastapi
uvicorn
pydantic
requests
psycopg2-binary
python-multipart
python-dotenv
google-generativeai>=0.8.3